pyproject.toml
[project]
name = "supply-chain-demo"
version = "0.1.0"
description = "End-to-End Supply Chain Control Tower — RisingWave + Agentic AI"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
